source/requirements.txt
flask
sympy
numpy
python-dotenv
requests
pyjwt